Add files via upload
This is a responsive and modern website I created for my final project in the Web Engineering subject. It is designed for a digital marketing agency and includes multiple pages like Home, About Us, Sign In, and Log In. The website is fully responsive, meaning it works well on both mobile and desktop screens. It includes form validation (like checking email format and matching passwords), a clean layout, and smooth user interaction. I used HTML for structure, CSS for design, and JavaScript for form checks and interactivity. This project highlights my front-end web development skills.